+++ lib/Serverless.js
@@ -14,13 +14,13 @@ const Utils = require('./classes/Utils');
 const Service = require('./classes/Service');
 const Variables = require('./classes/Variables');
 const ConfigSchemaHandler = require('./classes/ConfigSchemaHandler');
-const EnvLoader = require('./classes/EnvLoader');
 const ServerlessError = require('./classes/Error').ServerlessError;
 const Version = require('./../package.json').version;
 const isStandaloneExecutable = require('./utils/isStandaloneExecutable');
 const resolveCliInput = require('./utils/resolveCliInput');
 const logDeprecation = require('./utils/logDeprecation');
 const eventuallyUpdate = require('./utils/eventuallyUpdate');
+const loadEnv = require('./loadEnv');
 
 class Serverless {
   constructor(config) {
@@ -37,7 +37,6 @@ class Serverless {
     this.variables = new Variables(this);
     this.pluginManager = new PluginManager(this);
     this.configSchemaHandler = new ConfigSchemaHandler(this);
-    this.envLoader = new EnvLoader(this);
 
     // use the servicePath from the options or try to find it in the CWD
     this.cliInputArgv = process.argv.slice(2);
@@ -56,7 +55,6 @@ class Serverless {
     this.classes.Error = ServerlessError;
     this.classes.PluginManager = PluginManager;
     this.classes.ConfigSchemaHandler = ConfigSchemaHandler;
-    this.classes.EnvLoader = EnvLoader;
 
     this.serverlessDirPath = path.join(os.homedir(), '.serverless');
     this.isStandaloneExecutable = isStandaloneExecutable;
@@ -85,10 +83,9 @@ class Serverless {
         // set the options and commands which were processed by the CLI
         this.pluginManager.setCliOptions(this.processedInput.options);
         this.pluginManager.setCliCommands(this.processedInput.commands);
-        this.envLoader.load();
 
-        return this.service
-          .load(this.processedInput.options)
+        return loadEnv(this)
+          .then(() => this.service.load(this.processedInput.options))
           .then(() => {
             // load all plugins
             return this.pluginManager.loadAllPlugins(this.service.plugins);
