+++ lib/loadEnv.test.js
@@ -1,12 +1,12 @@
 'use strict';
 
 const chai = require('chai');
-const runServerless = require('../../test/utils/run-serverless');
+const runServerless = require('../test/utils/run-serverless');
 
 chai.use(require('chai-as-promised'));
 const expect = require('chai').expect;
 
-describe('EnvLoader', () => {
+describe('loadEnv', () => {
   describe('with useDotenv', () => {
     it('should load matching stage env file if present', async () => {
       const result = await runServerless({
